refactor: hoist parseHeading to markdown-sections (Rule of Three)
parseHeading was an exact 15-line clone across scenario-coverage.ts and
test-skeleton.ts (jscpd-flagged). Hoist it into markdown-sections.ts — the
documented home for shared markdown section-walk primitives (computeSkipMask,
stripInlineComments) — and import in both. Canonical version uses /^\s/ (space
or tab, per CommonMark), behavior-preserving for all real headings. Full suite
2548 pass; lint clean; jscpd clone resolved.
